move widget code to widget.rs file
