+++ src/config.rs
@@ -1,6 +1,7 @@
 use std::{fs, path::PathBuf};
-use ratatui::layout::Flex;
 use serde::Deserialize;
+
+use crate::widget::{ContainerFlex, Direction, FumWidget, LabelAlignment};
 
 #[derive(Debug, Clone, Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -37,54 +38,6 @@ impl Align {
 }
 
 #[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum Direction {
-    Vertical,
-    Horizontal
-}
-
-impl Direction {
-    pub fn to_dir(&self) -> ratatui::layout::Direction {
-        match self {
-            Self::Horizontal => ratatui::layout::Direction::Horizontal,
-            Self::Vertical => ratatui::layout::Direction::Vertical
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum LabelAlignment {
-    Left,
-    Center,
-    Right
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ContainerFlex {
-    Start,
-    Center,
-    End,
-    #[serde(rename = "space-around")]
-    SpaceAround,
-    #[serde(rename = "space-between")]
-    SpaceBetween
-}
-
-impl ContainerFlex {
-    pub fn to_flex(&self) -> Flex {
-        match self {
-            Self::Start         => Flex::Start,
-            Self::Center        => Flex::Center,
-            Self::End           => Flex::End,
-            Self::SpaceAround   => Flex::SpaceAround,
-            Self::SpaceBetween  => Flex::SpaceBetween
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
 pub struct Config {
     #[serde(default = "players")]
     pub players: Vec<String>,
@@ -112,33 +65,6 @@ pub struct Config {
 
     #[serde(default = "layout")]
     pub layout: Vec<FumWidget>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "type")]
-#[serde(rename_all = "lowercase")]
-pub enum FumWidget {
-    Container {
-        width: u16,
-        height: u16,
-        direction: Direction,
-        children: Vec<FumWidget>,
-        flex: Option<ContainerFlex>,
-    },
-    #[serde(rename = "cover-art")]
-    CoverArt {
-        width: u16,
-        height: u16
-    },
-    Label {
-        text: String,
-        align: Option<LabelAlignment>
-    },
-    Button {
-        text: String,
-        action: String,
-        exec: Option<String>
-    }
 }
 
 impl Default for Config {
