+++ src/widget.rs
@@ -1,0 +1,76 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Direction {
+    Vertical,
+    Horizontal
+}
+
+impl Direction {
+    pub fn to_dir(&self) -> ratatui::layout::Direction {
+        match self {
+            Self::Horizontal => ratatui::layout::Direction::Horizontal,
+            Self::Vertical => ratatui::layout::Direction::Vertical
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum LabelAlignment {
+    Left,
+    Center,
+    Right
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum ContainerFlex {
+    Start,
+    Center,
+    End,
+    #[serde(rename = "space-around")]
+    SpaceAround,
+    #[serde(rename = "space-between")]
+    SpaceBetween
+}
+
+impl ContainerFlex {
+    pub fn to_flex(&self) -> ratatui::layout::Flex {
+        match self {
+            Self::Start         => ratatui::layout::Flex::Start,
+            Self::Center        => ratatui::layout::Flex::Center,
+            Self::End           => ratatui::layout::Flex::End,
+            Self::SpaceAround   => ratatui::layout::Flex::SpaceAround,
+            Self::SpaceBetween  => ratatui::layout::Flex::SpaceBetween
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize)]
+#[serde(tag = "type")]
+#[serde(rename_all = "lowercase")]
+pub enum FumWidget {
+    Container {
+        width: u16,
+        height: u16,
+        direction: Direction,
+        children: Vec<FumWidget>,
+        flex: Option<ContainerFlex>,
+    },
+    #[serde(rename = "cover-art")]
+    CoverArt {
+        width: u16,
+        height: u16
+    },
+    Label {
+        text: String,
+        align: Option<LabelAlignment>
+    },
+    Button {
+        text: String,
+        action: String,
+        exec: Option<String>
+    }
+}
